requirements/tests.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements/tests.in
#
black==23.10.0
    # via -r requirements/tests.in
cachetools==5.3.1
    # via tox
chardet==5.2.0
    # via tox
click==8.1.7
    # via black
colorama==0.4.6
    # via tox
coverage[toml]==7.3.2
    # via
    #   coverage
    #   pytest-cov
distlib==0.3.7
    # via virtualenv
exceptiongroup==1.1.3
    # via pytest
filelock==3.12.4
    # via
    #   tox
    #   virtualenv
flake8==6.1.0
    # via -r requirements/tests.in
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements/tests.in
mccabe==0.7.0
    # via flake8
mypy==1.6.1
    # via -r requirements/tests.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.2
    # via
    #   black
    #   pyproject-api
    #   pytest
    #   tox
pathspec==0.11.2
    # via black
platformdirs==3.11.0
    # via
    #   black
    #   tox
    #   virtualenv
pluggy==1.3.0
    # via
    #   pytest
    #   tox
pycodestyle==2.11.1
    # via flake8
pyflakes==3.1.0
    # via flake8
pyproject-api==1.6.1
    # via tox
pytest==7.4.2
    # via
    #   -r requirements/tests.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements/tests.in
pytest-django==4.5.2
    # via -r requirements/tests.in
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   mypy
    #   pyproject-api
    #   pytest
    #   tox
tox==4.11.3
    # via
    #   -r requirements/tests.in
    #   tox-pyenv
tox-pyenv==1.1.0
    # via -r requirements/tests.in
typing-extensions==4.8.0
    # via
    #   black
    #   mypy
virtualenv==20.24.5
    # via tox
